+++ gauguin-core/src/main/kotlin/org/piepmeyer/gauguin/grid/Grid.kt
@@ -74,8 +74,6 @@ class Grid(
 
     fun isSolved(): Boolean = !cells.any { !it.isUserValueCorrect }
 
-    fun isNishioSolution(): Boolean = cells.all { it.isUserValueCorrect || (it.possibles.size == 1 && it.possibles.first() == it.value) }
-
     fun numberOfMistakes(): Int {
         logger.trace { "Calculating number of mistakes of:" }
         logger.trace { detailedToString() }
@@ -315,5 +313,7 @@ class Grid(
         return newGrid
     }
 
-    fun isNishioCheckable(): Boolean = cells.all { it.isUserValueSet || it.possibles.size == 1 } && !isSolved()
+    fun eachColumnContainsEachPossibleValue(): Boolean = gridSize.height == gridSize.largestSide()
+
+    fun eachRowContainsEachPossibleValue(): Boolean = gridSize.width == gridSize.largestSide()
 }