+++ gauguin-core/src/main/kotlin/org/piepmeyer/gauguin/grid/GridNishioLogic.kt
@@ -1,0 +1,83 @@
+package org.piepmeyer.gauguin.grid
+
+class GridNishioLogic(
+    private val grid: Grid,
+) {
+    fun isNishioSolution(): Boolean =
+        grid.cells.all { cell ->
+            cell.isUserValueCorrect ||
+                (cell.possibles.size == 1 && cell.possibles.first() == cell.value) ||
+                (
+                    cell.possibles.isEmpty() &&
+                        (
+                            (
+                                grid.eachRowContainsEachPossibleValue() &&
+                                    grid.getCellsAtSameRow(cell).all { it.isUserValueSet }
+                            ) ||
+                                (
+                                    grid.eachColumnContainsEachPossibleValue() &&
+                                        grid.getCellsAtSameColumn(cell).all { it.isUserValueSet }
+                                )
+                        )
+                )
+        }
+
+    fun isNishioCheckable(): Boolean =
+        grid.cells.all { cell ->
+            cell.isUserValueSet ||
+                cell.possibles.size == 1 ||
+                (
+                    cell.possibles.isEmpty() &&
+                        (
+                            (
+                                grid.eachRowContainsEachPossibleValue() &&
+                                    grid.getCellsAtSameRow(cell).all { it.isUserValueSet }
+                            ) ||
+                                (
+                                    grid.eachColumnContainsEachPossibleValue() &&
+                                        grid.getCellsAtSameColumn(cell).all { it.isUserValueSet }
+                                )
+                        )
+                )
+        } &&
+            !grid.isSolved()
+
+    fun solveViaNishioSolution() {
+        // first set all cells which have no possible set but are the only cell in row or column to have no value
+        grid.cells
+            .filter { !it.isUserValueSet && it.possibles.isEmpty() }
+            .forEach { cell ->
+                cell.setUserValueExtern(userValueFromFilledRowOrColumn(cell))
+            }
+
+        // the fill all cells with exactly one possible
+        grid.cells
+            .filter { !it.isUserValueSet && it.possibles.size == 1 }
+            .forEach { cell ->
+                cell.setUserValueExtern(cell.possibles.first())
+            }
+    }
+
+    private fun userValueFromFilledRowOrColumn(cell: GridCell): Int? {
+        val otherCellsOfRowOrColumn = mutableListOf<List<GridCell>>()
+
+        if (grid.eachColumnContainsEachPossibleValue()) {
+            otherCellsOfRowOrColumn += grid.getCellsAtSameColumn(cell)
+        }
+
+        if (grid.eachRowContainsEachPossibleValue()) {
+            otherCellsOfRowOrColumn += grid.getCellsAtSameRow(cell)
+        }
+
+        otherCellsOfRowOrColumn.forEach { otherCells ->
+            if (otherCells.all { it.isUserValueSet }) {
+                val possiblesLeftOver =
+                    grid.variant.possibleDigits - otherCells.map { it.userValue!! }
+
+                return possiblesLeftOver.first()
+            }
+        }
+
+        return null
+    }
+}